Add logs for VotingSession creation

+++ src/main/java/com/eduardozluhan/assembly/controller/request/OpenVotingSessionRequest.java
@@ -9,7 +9,7 @@ import java.time.LocalDateTime;
 @Value
 public class OpenVotingSessionRequest {
     @NotNull
-    Integer subjectId;
+    Long subjectId;
     @DateTimeFormat(iso = DateTimeFormat.ISO.DATE)
     LocalDateTime endsAt;
 }
